Rename types/index.d => types/server.d, other changes
Rename types/index.d.ts => types/server.d.ts
- done for clarity, need to update client verson of it as well
- relevant imports updated as well

Update certain type aliases in server.d.ts to interfaces
- better matches with what an interface/ type alias is
  - basically an interface only express a product type/ the shape of a
  ... thing
    - product type being a compounded type in a structure
  - a type alias on the other hand expresses the specific type of a thing
