+++ routes/middleware/userAuthMiddleware.ts
@@ -1,5 +1,5 @@
 import { RequestHandler, ErrorRequestHandler } from 'express';
-import { responseObj } from '../../types/index';
+import { responseObj } from '../../types/server';
 
 const postLogin: RequestHandler = (req, res, next): any => {
   const { username, _id } = req.user;
